Change SI options to array of hashes
Extensions might need a persistent UID to manage their data cache, but the displayed name may be different. Transform the returned 2D array into an array of hashes with uid/display keys for client consumption.

Also removes option retrieval from the stored records: Without a return statement, it would be overwritten anyway, and options that are not already cached wouldn’t be included.

Signed-off-by: Tobias Grasse <tg@glancr.de>