Replace setting “city” with “privacyConsent”
City was a core setting in mirr.OS 0.x, since it was used by several default widgets. Extensions in mirr.OS one should not rely on special cases. Another problem with this approach is that different data services might need different encodings of location information, which makes this setting useless for some widgets and potentially confuses the user.