Prevent errors on connection reset
Do not attempt to reset the network connection if the active connection is LAN (since we don’t have to clean up connections). Catch nmcli exit status 10 as well, since the connection might not exist for some other reason. Leaving a connection present does not hurt, but might conflict with later re-enabling.

Signed-off-by: Tobias Grasse <tg@glancr.de>