Move AP open/close logic from run_setup to Network.connect
Settings app displays the normal backend if just the network connection is failing, so that users don’t have to tap through the whole setup again. SettingExecution::Network.connect does not deal with the AP, which if open blocks any WiFi operations. This moves the required logic inside the host-agnostic superclass, so that calls to connect are always run without the AP open.

Signed-off-by: Tobias Grasse <tg@glancr.de>