Switch sub-method logic for switch to LAN
previously, this implied that resetting the WiFi connection only happens on a system reset, in which case LAN connections should be disabled (to prevent problems with DNS and to force setup to happen via WiFi). But if LAN is enabled as a connection method, we should delete existing connections in network-manager to prevent starting up stale connection profiles.

Signed-off-by: Tobias Grasse <tg@glancr.de>