Add method request_wifi_scan to NM::Commands
Triggers a scan for available networks on the primary WiFi device. When passed a SSID name, the scan will look for this SSID specifically. Requesting a scan while the device is already scanning would trigger an error from DBus; this is handled right in the method since the method’s goal is achieved anyway.

Signed-off-by: Tobias Grasse <tg@glancr.de>