Improve WiFi scan logic
Since we can’t check the LastScan property (not implemented in available snap versions of NM), we need to listen to AccessPointAdded signals. This implementation uses a separate thread to run the DBus signal listener loop, which terminates once a matching SSID has been found. In parallel, the main thread runs a while loop that checks the signal loop’s output each 2 seconds and terminates the whole process if no AP was found after 20 seconds. This does not take into account that the AP is already known to NM, so only use this when the AP cannot be found with ap_object_path.

Signed-off-by: Tobias Grasse <tg@glancr.de>