[Fix] Ensure inactive connection state on startup
If a connection was previously active and is now inactive on startup, this wasn’t reflected in the DB as persist_inactive_connection was changed to not mess with the `active` and `active_connection_path` fields. This ensures that the current state for all connections is properly synced on startup.

Signed-off-by: Tobias Grasse <tg@glancr.de>