[Fix] Open AP immediately on connectivity issues
If the network connection was saved in NM, but the device can’t get online in the test period, the setup aborts without an active AP. This is the case if a user entered the wrong WiFi password. AP will be scheduled to open in 15min, but that’s too long in case of a botched setup. This change ensures the AP will be opened immediately after any exceptions during setup.

Signed-off-by: Tobias Grasse <tg@glancr.de>