fix: Avoid duplicate file conflict in source generator
We don’t want to confuse extension developers with file conflicts during a regular generator run. The default (empty) main class is provided by the base plugin generator, so we need to remove it. Maybe also a good hint to think about refactoring the source scaffolding with a separate hooks class file.

Signed-off-by: Tobias Grasse <tg@glancr.de>