fix: Return clear error on extension mounting failures
When the database contains an extension with no corresponding Engine class, the application crashes with a rather cryptic «undefined method `config` for nil». I opted to raise here instead of ignoring the error, since it might lead to subtle but fatal bugs further down. If an extension exists in the DB but does not have their gem installed, the user either tampered with their installation or something larger is broken and they should contact support.

Signed-off-by: Tobias Grasse <tg@glancr.de>