refactor: SettingExecution::System.timezone optimization
This method currently only does meaningful work on Linux hosts, so we should check that condition first. On all other platforms, log a warning instead of raising to prevent confusing errors. This is in line with other parts of the system that will fail if running on non-Linux hosts, which is an unsupported use case at the moment. Development envs are treated differently anyway.

Signed-off-by: Tobias Grasse <tg@glancr.de>