fix: Only schedule reboot once system time is correct
Without this check, the reboot is scheduled on first boot while the system time is still set to the last known timestamp (image creation). This would trigger a sudden reboot during setup once the system time is correct, because that will be much later and thus trigger the reboot schedule.

Signed-off-by: Tobias Grasse <tg@glancr.de>