refactor: Use StateCache as Redis wrapper
Instead of a SIngleton that’s only reliable inside the main thread, we persist the initial state retrieval and all successive state updates in namespaced Redis keys. The massive switch statement for initialization is not ideal, but keeps the class self-contained until the whole state logic is rewritten to push a single, clear state to clients.

Signed-off-by: Tobias Grasse <tg@glancr.de>